--- permbus-api/build.gradle
@@ -1,37 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'maven-publish'
-}
-
-android {
-    namespace 'com.example.permbus'
-    compileSdk 34
-
-    defaultConfig {
-        minSdk 21
-    }
-
-    publishing {
-        singleVariant("release")
-    }
-}
-
-dependencies {
-}
-
-publishing {
-    publications {
-        release(MavenPublication) {
-            groupId = 'com.example'
-            artifactId = 'permbus-api'
-            version = '1.0'
-        }
-    }
-    repositories {
-        mavenLocal()
-    }
-}
-
-afterEvaluate {
-    publishing.publications.release.from(components["release"])
-}
