--- permbus-api/src/main/aidl/com/example/permbus/IPermissionPortal.aidl
@@ -1,7 +1,0 @@
-package com.example.permbus;
-
-import com.example.permbus.PermInfo;
-
-interface IPermissionPortal {
-    List<PermInfo> getPendingPermissions();
-}
