--- permbus-api/src/main/aidl/com/example/permbus/PermInfo.aidl
@@ -1,3 +1,0 @@
-package com.example.permbus;
-
-parcelable PermInfo;
