+++ permissionFlow/build.gradle
@@ -4,7 +4,7 @@ plugins {
 }
 
 android {
-    namespace 'com.example.permbus'
+    namespace 'com.example.permissionflow'
     compileSdk 34
 
     defaultConfig {
@@ -17,7 +17,6 @@ android {
 }
 
 dependencies {
-    api project(':permbus-api')
     implementation 'androidx.core:core:1.12.0'
 }
 
@@ -25,7 +24,7 @@ publishing {
     publications {
         release(MavenPublication) {
             groupId = 'com.example'
-            artifactId = 'permbus-ui'
+            artifactId = 'permission-flow'
             version = '1.0'
         }
     }
