+++ permissionFlow/src/main/aidl/com/example/permissionflow/IPermissionPortal.aidl
@@ -1,0 +1,7 @@
+package com.example.permissionflow;
+
+import com.example.permissionflow.PermInfo;
+
+interface IPermissionPortal {
+    List<PermInfo> getPendingPermissions();
+}
