+++ permissionFlow/src/main/aidl/com/example/permissionflow/PermInfo.aidl
@@ -1,0 +1,3 @@
+package com.example.permissionflow;
+
+parcelable PermInfo;
